+++ NEWS.md
@@ -1,4 +1,6 @@
-# tidypolars (development version)
+# tidypolars 0.14.1
+
+* `tidypolars` requires `polars` >= 1.1.0 (#222).
 
 ## Bug fixes